pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "calibration"
version = "0.1.0"
description = "UI-only shell for rebuilding the calibration project."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "opencv-python",
    "pyorbbecsdk-community",
    "PySide6>=6.0",
    "numpy<2",
    "pyrealsense2>=2.55.1.6486",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1",
]

[project.scripts]
calibration = "calibration.app:main"

[tool.setuptools.packages.find]
where = ["src"]
